refactor: remove needless type casts

+++ qreki.go
@@ -157,8 +157,8 @@ func calcEclipticLongitude(sl float64, longitude float64) float64 {
 }
 
 func getNishinibun(julian float64) (float64, float64, float64) {
-	julianIntegerPart := math.Floor(float64(julian))
-	julianDecimalPart := float64(julian) - float64(julianIntegerPart) + tz
+	julianIntegerPart := math.Floor(julian)
+	julianDecimalPart := julian - julianIntegerPart + tz
 	nnj := (julianDecimalPart+0.5)/36525.0 + (julianIntegerPart-2451545.0)/36525.0
 	return julianIntegerPart, julianDecimalPart, nnj
 }
